src/apps/nfc/session/resolvers/insert-sessions.resolver.spec.ts
import { Test, TestingModule } from '@nestjs/testing';

// custom items
import { InsertSessionsResolver } from './insert-sessions.resolver';
import { ICommandBus } from '@hades/shared/domain/bus/command-bus';
import { IQueryBus } from '@hades/shared/domain/bus/query-bus';
import { sessions } from '@hades/nfc/session/infrastructure/seeds/session.seed'

describe('InsertSessionsResolver', () => 
{
    let resolver: InsertSessionsResolver;
    let queryBus: IQueryBus;
    let commandBus: ICommandBus;

    beforeEach(async () => 
    {
        const module: TestingModule = await Test.createTestingModule({
            providers: [
                InsertSessionsResolver,
                {
                    provide: IQueryBus,
                    useValue: {
                        ask: () => {},
                    }
                },
                {
                    provide: ICommandBus,
                    useValue: {
                        dispatch: () => {},
                    }
                },
            ]
        }).compile();

        resolver    = module.get<InsertSessionsResolver>(InsertSessionsResolver);
        queryBus    = module.get<IQueryBus>(IQueryBus);
        commandBus  = module.get<ICommandBus>(ICommandBus);
    });

    it('InsertSessionsResolver should be defined', () => 
    {
        expect(resolver).toBeDefined();
    });

    // Test get method
    describe('main', () => 
    {
        it('InsertSessionsResolver should be defined', () => 
        {
            expect(resolver).toBeDefined();
        });

        it('should return an sessions created', async () => 
        {
            expect(await resolver.main(sessions)).toBe(true);
        });
    });
});